Triangle Rendering
Added functionality to be able to render triangles. Have used an intersection technique similar to the one I used in my Cloth simulator. I calculate u,v coordinates by calculating the barycentric coordinates of the hit point and weighting that with the vertices' u,v coords (haven't tested this thoroughly yet). Added bounding boxes to the triangles such that render times speed up dramatically.
For the future I should remember to keep t within t_min and t_max so that I'm not smacking my head on the table wondering why the triangle is black.