BVH Speed Up
Removed code that duplicated the objects array in the BVH Nodes. I found that sorting in place literally has the same result, and it is soo much faster.

BVH construction is still rather slow for very complex objects. For example, the common dragon.obj has ~900,000 faces and a BVH can be constructed in about 8 minutes (before making the optimisation, the BVH would have taken hours to construct).